+++ src/Castle.Facilities.AspNet.SystemWeb/Lifestyle/PerWebRequestAttribute.cs
@@ -1,4 +1,4 @@
-// Copyright 2004-2012 Castle Project - http://www.castleproject.org/
+// Copyright 2004-2011 Castle Project - http://www.castleproject.org/
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,30 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-
-#if FEATURE_SYSTEM_WEB
-
-namespace Castle.MicroKernel.Lifestyle
+namespace Castle.Facilities.AspNet.SystemWeb.Lifestyle
 {
-	using Castle.MicroKernel.Context;
-	using Castle.MicroKernel.Lifestyle.Scoped;
+	using System;
 
-	public class WebRequestScopeAccessor : IScopeAccessor
+	using Castle.Core;
+
+	[AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
+	public sealed class PerWebRequestAttribute : ScopedAttribute
 	{
-		public void Dispose()
+		public PerWebRequestAttribute()
 		{
-			var scope = PerWebRequestLifestyleModule.YieldScope();
-			if (scope != null)
-			{
-				scope.Dispose();
-			}
-		}
-
-		public ILifetimeScope GetScope(CreationContext context)
-		{
-			return PerWebRequestLifestyleModule.GetScope();
+			ScopeAccessorType = typeof(WebRequestScopeAccessor);
 		}
 	}
 }
-
-#endif