+++ src/Castle.Windsor/MicroKernel/Registration/BasedOnDescriptor.cs
@@ -353,17 +353,6 @@ namespace Castle.MicroKernel.Registration
 			return Configure(c => c.LifestyleBoundToNearest<TBaseForRoot>());
 		}
 
-#if FEATURE_SYSTEM_WEB
-		/// <summary>
-		///   Sets component lifestyle to instance per web request.
-		/// </summary>
-		/// <returns> </returns>
-		public BasedOnDescriptor LifestylePerWebRequest()
-		{
-			return Configure(c => c.LifestylePerWebRequest());
-		}
-#endif
-
 		/// <summary>
 		///   Sets component lifestyle to pooled. If <paramref name = "initialSize" /> or <paramref name = "maxSize" /> are not set default values will be used.
 		/// </summary>